Improved items forms. Refactored Contact form. Got rid of useless presentations of subtypes. Minor other changes.
